Rework
- SocketTestRunner uses a Blocker for reading streams
- Add description to mutant result (and implement in Error result)
- Move some mapping logic
- ErrorDuringTestRun results in Killed mutant
- 2 retries instead of only 1
